pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eripotter-common"
version = "0.1.4"
description = "Common utilities for EriPotter microservices"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "EriPotter Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "asyncpg>=0.28.0",  # async database support
    "PyJWT>=2.8.0",     # JWT token support
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://eripotter.com"
Repository = "https://github.com/eripotter/eripotter-common"

[tool.hatch.build.targets.wheel]
packages = ["eripotter_common"]

[tool.pytest.ini_options]
addopts = "-v --cov=eripotter_common --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["eripotter_common"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true